Bugfix: 'at' got turned off when loading with Python 2.3.
git-svn-id: https://rox.svn.sourceforge.net/svnroot/rox/trunk/Memo/trunk/Memo@3040 66de3db3-b00d-0410-b41b-f4738ad19bea

+++ Window.py
@@ -185,4 +185,4 @@ class Window(g.Window):
 		if delay > 60 * 60 * 24:
 			delay = 60 * 60 * 24
 
-		self.timeout = g.timeout_add(1000 * delay, self.timeout_cb)
+		self.timeout = g.timeout_add(int(1000 * delay), self.timeout_cb)
